pyproject.toml
[project]
name = "lewaf"
version = "0.7.0"
description = "A Web Application Firewall for Python that stops attacks before they reach your application code, with comprehensive audit logging and compliance support."
readme = "README.md"
authors = [
    { name = "Stefane Fermigier", email = "sf@abilian.com" }
]
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
keywords = ["waf", "security", "modsecurity", "firewall", "web-security", "owasp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Middleware",
]
dependencies = [
    "starlette>=0.41.0",
    "httpx>=0.27.0",
    "uvicorn>=0.32.0",
    "redis>=7.0.1",
    "pyyaml>=6.0.3",
    "click>=8.2.1",
]

[project.urls]
Homepage = "https://github.com/abilian/lewaf"
Documentation = "https://github.com/abilian/lewaf/tree/main/docs"
Repository = "https://github.com/abilian/lewaf"
Issues = "https://github.com/abilian/lewaf/issues"

[project.scripts]
lewaf = "lewaf:main"
lewaf-proxy = "lewaf.proxy.cli:main"
lewaf-validate = "lewaf.cli.validate:validate"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    # Test
    "pytest>=8.4.2",
    "nox>=2025.11.12",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    # Lint / Check
    "isort>=6.0.1",
    "ruff>=0.12.12",
    "pyrefly>=0.32.0",
    "ty>=0.0.1a25",
    "monkeytype>=23.3.0",
    "mypy>=1.19.0",
    "typeguard>=4.4.4",
    "types-pyyaml>=6.0.12.20250915",
    # Other
    "httpx>=0.28.1",
    "starlette>=0.47.3",
    "uvicorn>=0.35.0",
]
examples = [
    # Web frameworks for integration examples
    "flask>=3.1.0",
    "fastapi>=0.115.0",
    "django>=5.1.0",
    # WSGI/ASGI servers
    "uvicorn>=0.35.0",
    "gunicorn>=23.0.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.27.0",
    "mike>=2.1.0",
]

#
# QA tools
#

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
check_untyped_defs = true
